+++ StreamMaster.Streams/Buffers/CircularRingBuffer.cs
@@ -14,13 +14,13 @@ public sealed partial class CircularRingBuffer : ICircularRingBuffer
 
     public long GetNextReadIndex()
     {
-        return _writeBytes;
+        return WriteBytes;
     }
 
     private int CalculateClientReadIndex(long readByteIndex)
     {
         // Calculate the byte difference
-        long byteDifference = _writeBytes - readByteIndex;
+        long byteDifference = WriteBytes - readByteIndex;
 
         // Convert byte difference to buffer index difference
         int indexDifference = (int)(byteDifference % _buffer.Length);
@@ -37,16 +37,11 @@ public sealed partial class CircularRingBuffer : ICircularRingBuffer
         return clientReadIndex;
     }
 
-
-
     public async Task<int> ReadChunkMemory(long readIndex, Memory<byte> target, CancellationToken cancellationToken)
     {
         Guid correlationId = Guid.NewGuid();
 
         Stopwatch stopwatch = Stopwatch.StartNew();
-
-        //PerformanceBpsMetrics metrics = _performanceMetrics.GetOrAdd(ClientId, key => new PerformanceBpsMetrics());
-
         CancellationToken linkedToken = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken, StopVideoStreamingToken.Token).Token;
 
         int bytesRead = 0;
@@ -83,32 +78,9 @@ public sealed partial class CircularRingBuffer : ICircularRingBuffer
             clientReadIndex = (clientReadIndex + bytesToRead) % bufferLength;
         }
 
-        //metrics.RecordBytesProcessed(bytesRead);
-        //_statisticsManager.AddBytesRead(ClientId, bytesRead);
-
-        //if (_clientReadIndexes.TryGetValue(ClientId, out int value))
-        //{
-        //    // Enhanced structured logging
-        //    var logData = new
-        //    {
-        //        Event = "ReadChunkMemory",
-        //        CorrelationId = correlationId,
-        //        ClientId,
-        //        VideoStreamName,
-        //        ClientReadIndex = value,
-        //        BytesRead = bytesRead,
-        //        ReadDurationMs = stopwatch.ElapsedMilliseconds,
-        //        BufferOccupancy = CalculateBufferOccupancy(),
-        //        ActiveReaders = _clientReadIndexes.Count,
-        //        DistanceToOldestReader = CalculateDistanceToOldestReader(correlationId)
-        //    };
-        //    _readLogger.LogDebug(JsonSerializer.Serialize(logData));
-        //}
-
         stopwatch.Stop();
         return bytesRead;
     }
-
 
     public async Task<int> WriteChunk(Memory<byte> data, CancellationToken cancellationToken)
     {
@@ -142,7 +114,7 @@ public sealed partial class CircularRingBuffer : ICircularRingBuffer
                 data[..lengthToWrite].CopyTo(bufferSlice);
 
                 _writeIndex = (_writeIndex + lengthToWrite) % _buffer.Length;
-                _writeBytes += lengthToWrite;
+                WriteBytes += lengthToWrite;
                 _bytesWritten += lengthToWrite;
 
                 data = data[lengthToWrite..];
@@ -187,7 +159,6 @@ public sealed partial class CircularRingBuffer : ICircularRingBuffer
         }
     }
 
-
     /// <summary>
     /// Signals all waiting readers that new data has been written.
     /// </summary>
@@ -195,260 +166,5 @@ public sealed partial class CircularRingBuffer : ICircularRingBuffer
     {
         SignalReaders();
     }
-
-    /// <summary>
-    /// Calculates the amount of data currently stored in the buffer.
-    /// This method determines how much of the buffer is occupied with data
-    /// that has been written but not necessarily read by all clients.
-    /// </summary>
-    /// <returns>The total number of bytes occupied in the buffer.</returns>
-    //private int CalculateBufferOccupancy()
-    //{
-    //    // If the buffer hasn't flipped (i.e., write index hasn't reached the end of the buffer),
-    //    // occupancy is simply the current position of the write index.
-    //    if (!HasBufferFlipped)
-    //    {
-    //        return _writeIndex;
-    //    }
-
-    //    // When the buffer has flipped, it means the write operation has wrapped around to the start
-    //    // and may be overwriting older data. In this case, the method needs to find the oldest 
-    //    // (furthest behind) read index among all clients. This index helps in understanding
-    //    // the part of the buffer that has data which might not have been read yet.
-    //    int oldestReadIndex = _clientReadIndexes.Values.DefaultIfEmpty(_buffer.Length).Min();
-
-    //    // If there are no readers, the whole buffer is considered occupied.
-    //    if (oldestReadIndex == _buffer.Length)
-    //    {
-    //        return _buffer.Length;
-    //    }
-
-    //    // Calculate the occupied space based on the distance between the write index and the oldest read index.
-    //    // This helps determine how much new data can be accepted before overwriting unread data.
-    //    return _writeIndex >= oldestReadIndex ?
-    //        _writeIndex - oldestReadIndex : // Write index is ahead of the oldest read index
-    //        _buffer.Length - oldestReadIndex + _writeIndex; // Write index has wrapped around past the oldest read index
-    //}
-
-
-
-    /// <summary>
-    /// Calculates the maximum distance from the write index to any read index.
-    /// This represents the furthest (oldest) unread data in the buffer.
-    /// </summary>
-    /// <param name="correlationId">Correlation identifier for logging purposes.</param>
-    /// <returns>The maximum distance to the oldest reader in the buffer.</returns>
-    //private int CalculateDistanceToOldestReader(Guid correlationId)
-    //{
-    //    int maxDistance = 0;
-    //    foreach (KeyValuePair<Guid, int> clientReadIndexEntry in _clientReadIndexes)
-    //    {
-    //        int distance = CalculateDistance(clientReadIndexEntry.Value, correlationId);
-    //        if (distance > maxDistance)
-    //        {
-    //            maxDistance = distance;
-    //        }
-    //    }
-
-    //    // Structured logging
-    //    var logData = new
-    //    {
-    //        Event = "CalculateDistanceToOldestReader",
-    //        CorrelationId = correlationId,
-    //        VideoStreamName,
-    //        MaxDistanceToOldestReader = maxDistance
-    //    };
-    //    _distanceLogger.LogDebug(System.Text.Json.JsonSerializer.Serialize(logData));
-
-    //    return maxDistance;
-    //}
-
-    /// <summary>
-    /// Finds the closest read index to the write index, indicating the nearest reader.
-    /// </summary>
-    /// <param name="correlationId">Correlation identifier for logging purposes.</param>
-    /// <returns>The shortest distance to the nearest reader in the buffer.</returns>
-    //private int CalculateClosestReadIndexDistance(Guid correlationId)
-    //{
-    //    int closestDistance = _buffer.Length;
-
-    //    foreach (KeyValuePair<Guid, int> clientReadIndexEntry in _clientReadIndexes)
-    //    {
-    //        int distance = CalculateDistance(clientReadIndexEntry.Value, correlationId);
-    //        if (distance < closestDistance)
-    //        {
-    //            closestDistance = distance;
-    //        }
-    //    }
-
-    //    // Structured logging
-    //    var logData = new
-    //    {
-    //        Event = "CalculateClosestReadIndexDistance",
-    //        CorrelationId = correlationId,
-    //        ClosestDistance = closestDistance,
-    //        ActiveReaders = _clientReadIndexes.Count
-    //    };
-    //    _distanceLogger.LogDebug(JsonSerializer.Serialize(logData));
-
-    //    return closestDistance;
-    //}
-
-    /// <summary>
-    /// Calculates the distance between a given read index and the current write index.
-    /// This helps determine how far behind a reader is in the buffer.
-    /// </summary>
-    /// <param name="readIndex">The read index to calculate the distance from.</param>
-    /// <param name="correlationId">Correlation identifier for logging purposes.</param>
-    /// <returns>The calculated distance from the read index to the write index.</returns>
-    private int CalculateDistance(int readIndex, Guid correlationId)
-    {
-        if (!HasBufferFlipped)
-        {
-            // Log when the buffer hasn't flipped
-            LogCalculateDistance(correlationId, readIndex, _buffer.Length, "Buffer not flipped");
-            return _buffer.Length;
-        }
-
-        // Handle wrap-around case
-        int ret = _buffer.Length - _writeIndex + readIndex;
-
-        // Structured logging
-        LogCalculateDistance(correlationId, readIndex, ret, "Calculated distance");
-
-        return ret;
-    }
-
-
-
-
-    //private void CheckAndReportClientOverwrites(int lengthToWrite, Guid correlationId)
-    //{
-
-    //    int newWriteIndex = (_writeIndex + lengthToWrite) % _buffer.Length;
-    //    foreach (KeyValuePair<Guid, int> clientReadIndexEntry in _clientReadIndexes)
-    //    {
-    //        int clientReadIndex = clientReadIndexEntry.Value;
-
-    //        _overwriteLogger.LogDebug($"CheckAndReportClientOverwrites {_buffer.Length} {clientReadIndex} {_writeIndex} {newWriteIndex} {lengthToWrite}");
-
-    //        if (IsClientOverwritten(clientReadIndex, _writeIndex, newWriteIndex))
-    //        {
-    //            Guid clientId = clientReadIndexEntry.Key;
-    //            _overwriteLogger.LogWarning(System.Text.Json.JsonSerializer.Serialize(new
-    //            {
-    //                Event = "ClientOverwrite",
-    //                CorrelationId = correlationId,
-    //                ClientId = clientId,
-    //                VideoStreamName,
-    //                ClientReadIndex = clientReadIndexEntry.Value,
-    //                WriteIndex = _writeIndex,
-    //                NewWriteIndex = newWriteIndex
-    //            }));
-    //            _logger.LogWarning($"Client {clientId}'s read index {clientReadIndex} newWriteIndex {newWriteIndex} in {VideoStreamName} ");
-
-    //            _clientReadIndexes[clientId] = CalculateSafeReadIndex(newWriteIndex);
-    //            // ResizeBuffer();
-    //        }
-    //    }
-    //}
-
-    //private void CheckAndReportClientOverwrites(int lengthToWrite, Guid correlationId)
-    //{
-    //    int newWriteIndex = (_writeIndex + lengthToWrite) % _buffer.Length;
-
-    //    foreach (KeyValuePair<Guid, int> clientReadIndexEntry in _clientReadIndexes)
-    //    {
-    //        int clientReadIndex = clientReadIndexEntry.Value;
-
-    //        if (IsClientOverwritten(clientReadIndex, _writeIndex, newWriteIndex))
-    //        {
-    //            AdjustClientReadIndex(clientReadIndexEntry.Key, newWriteIndex, correlationId);
-    //        }
-    //    }
-    //}
-
-
-    /// <summary>
-    /// Adjusts the read index of a client to a safe position to avoid overwrites.
-    /// </summary>
-    //private void AdjustClientReadIndex(Guid clientId, int newWriteIndex, Guid correlationId)
-    //{
-    //    int safeReadIndex = CalculateSafeReadIndex(newWriteIndex);
-    //    _clientReadIndexes[clientId] = safeReadIndex;
-
-    //    // Log the adjustment for debugging and monitoring
-    //    _overwriteLogger.LogWarning(JsonSerializer.Serialize(new
-    //    {
-    //        Event = "AdjustClientReadIndex",
-    //        CorrelationId = correlationId,
-    //        ClientId = clientId,
-    //        SafeReadIndex = safeReadIndex,
-    //        OriginalReadIndex = _clientReadIndexes[clientId]
-    //    }));
-    //}
-
-    //private int CalculateThrottleWaitTime(Guid correlationId)
-    //{
-    //    int throttleThreshold = (int)(_buffer.Length * 0.10); // Threshold for throttling (10% of buffer size)
-    //    int closestReadIndexDistance = CalculateClosestReadIndexDistance(correlationId);
-
-    //    if (closestReadIndexDistance > throttleThreshold)
-    //    {
-    //        return 0; // No throttling required
-    //    }
-
-    //    if (closestReadIndexDistance == 0)
-    //    {
-    //        return maxDynamicWaitTimeMs; // Maximum wait if a client is immediately behind the write index
-    //    }
-
-    //    // Calculate proportional wait time based on proximity to the closest reader
-    //    double proximity = (double)closestReadIndexDistance / throttleThreshold;
-    //    int calculatedWaitTime = (int)(-maxDynamicWaitTimeMs * Math.Log(proximity));
-
-    //    // Log the calculated wait time for monitoring
-    //    LogDynamicWaitTime(correlationId, throttleThreshold, closestReadIndexDistance, calculatedWaitTime);
-
-    //    return Math.Min(calculatedWaitTime, maxDynamicWaitTimeMs); // Cap the wait time to a maximum value
-    //}
-
-
-    //private int CalculateSafeReadIndex(int newWriteIndex)
-    //{
-    //    int bufferLength = _buffer.Length;
-
-    //    newWriteIndex += ((int)(bufferLength * 0.10)) % bufferLength;
-
-    //    // Implement logic to calculate a safe read index for the client
-    //    // This might be the new write index, or some position before it, depending on your buffer's logic
-    //    return newWriteIndex;
-    //}
-
-    //private bool IsClientOverwritten(int clientReadIndex, int initialWriteIndex, int newWriteIndex)
-    //{
-    //    _overwriteLogger.LogDebug($"IsClientOverwritten {clientReadIndex} {initialWriteIndex} {newWriteIndex}");
-
-    //    if (clientReadIndex == initialWriteIndex) // Client is waiting for data
-    //    {
-    //        return false;
-    //    }
-
-
-    //    //Write doesnt wrap
-    //    if (initialWriteIndex <= newWriteIndex)
-    //    {
-    //        //In the middle of the write
-    //        bool test = clientReadIndex > initialWriteIndex && clientReadIndex <= newWriteIndex;
-    //        _overwriteLogger.LogDebug($"IsClientOverwritten No wrap {test}");
-    //        return test;
-    //    }
-    //    // Write wraps around
-
-    //    // Check if clientReadIndex is outside the range of newWriteIndex to initialWriteIndex
-    //    bool isOverwritten = !(clientReadIndex > newWriteIndex && clientReadIndex <= initialWriteIndex);
-    //    _overwriteLogger.LogDebug($"IsClientOverwritten wrap {isOverwritten}");
-    //    return isOverwritten;
-    //}
 
 }
