+++ StreamMaster.Streams/Factories/CircularRingBufferFactory.cs
@@ -5,10 +5,10 @@ using StreamMaster.Infrastructure.VideoStreamManager.Buffers;
 
 namespace StreamMaster.Streams.Factories;
 
-public sealed class CircularRingBufferFactory(IStatisticsManager statisticsManager, IInputStatisticsManager inputStatisticsManager, IMemoryCache memoryCache, ILoggerFactory loggerFactory) : ICircularRingBufferFactory
+public sealed class CircularRingBufferFactory(IInputStatisticsManager inputStatisticsManager, IMemoryCache memoryCache, ILoggerFactory loggerFactory) : ICircularRingBufferFactory
 {
     public ICircularRingBuffer CreateCircularRingBuffer(VideoStreamDto videoStreamDto, string ChannelId, string ChannelName, int rank)
     {
-        return new CircularRingBuffer(videoStreamDto, ChannelId, ChannelName, statisticsManager, inputStatisticsManager, memoryCache, rank, loggerFactory);
+        return new CircularRingBuffer(videoStreamDto, ChannelId, ChannelName, inputStatisticsManager, memoryCache, rank, loggerFactory);
     }
 }
